btrfs-progs: Fix partitioned loop devices resolve
When using partitions on a loop device, the device's name can be
/dev/loop0p1 or similar, and no relevant entry exists in the /sys
filesystem, so the current resolve_loop_device function fails.

This patch adds a fallback to using loopdev API which will get the
correct backing file even for partitioned loop devices, at the expense
of using more memory, hence using it as a fallback only.